Added support for Descriptors sets and texturing
The pipeline layout of a pipeline can now be specified using the PipelineLayoutInfo class. This will also generate descriptor pools which can be used to set uniform buffers, samplers or textures in the pipeline using a command buffer.
The command buffers can now be used to create textures from raw data or from a loaded file.