Added sponza model. Added StaticMesh and Scene classes.  Redid model loading to load the entire .gltf and allow for usage of specific meshes/materials/scenes. Added camera movement.
Added StaticMesh which serves as a mesh with a transform attached to it.
Added Scene class which is currently a collection of static meshes.
The model loader now stores GLTFResource objects. Each GLTFResource contains the meshes, materials, textures and scenes found in the file that are found in the .gltf file.  Destroying the GLTFResource will delete any resources in it which aren't being used elsewhere in the application.
The camera now has limited movement used during the testing using th WASDQE keys and the left and right arrows for rotation around the Y axis.